Code cleanup and happy clippy!

+++ orion-e2e-tests/src/lib.rs
@@ -61,7 +61,7 @@ pub fn cleanup_config_file(path: &std::path::Path) {
     }
 }
 
-pub fn parse_metric_value(prometheus_output: &str, metric_name: &str) -> Option<f64> {
+pub fn parse_metric_value(prometheus_output: &str, metric_name: &str) -> Option<u64> {
     for line in prometheus_output.lines() {
         if line.starts_with('#') {
             continue;
@@ -69,7 +69,7 @@ pub fn parse_metric_value(prometheus_output: &str, metric_name: &str) -> Option<
         if line.starts_with(metric_name) {
             let parts: Vec<&str> = line.split_whitespace().collect();
             if let Some(val_str) = parts.last() {
-                if let Ok(val) = val_str.parse::<f64>() {
+                if let Ok(val) = val_str.parse::<u64>() {
                     return Some(val);
                 }
             }
